+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.29.0] - 2020-08-31
 
+### Fixed
+- NumericRange: Incorrect 0 value handler.
+
 
 ## [0.28.0] - 2020-08-17